requirements.txt
asgiref==3.8.1
Django==5.0.7
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.0.8
sqlparse==0.5.1
tzdata==2024.1
